Fixes CVE-2014-3577
It was found that the fix for CVE-2012-6153 was incomplete: the code
added to
check that the server hostname matches the domain name in a subject's
Common
Name (CN) field in X.509 certificates was flawed. A man-in-the-middle
attacker
could use this flaw to spoof an SSL server using a specially crafted
X.509
certificate.
The fix for CVE-2012-6153 was intended to address the incomplete patch
for
CVE-2012-5783. This means the issue is now completely resolved by
applying
this patch and the 06_fix_CVE-2012-5783.patch.

References:

upstream announcement:
https://mail-archives.apache.org/mod_mbox/www-announce/201408.mbox/CVE-2014-3577

Fedora-Fix:
http://pkgs.fedoraproject.org/cgit/jakarta-commons-httpclient.git/tree/jakarta-commons-httpclient-CVE-2014-3577.patch

CentOS-Fix:
https://git.centos.org/blob/rpms!jakarta-commons-httpclient/SOURCES!jakarta-commons-httpclient-CVE-2014-3577.patch

Debian-Bug: https://bugs.debian.org/758086